% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_income_hale}
\alias{canada_income_hale}
\title{Canada income quintile HALE data}
\format{
A tibble with 5 rows and 9 variables analogous to
  \code{\link{england_imd_hale}} but with income-based stratification.
}
\source{
Statistics Canada, Health-Adjusted Life Expectancy by income
  quintile.
}
\usage{
canada_income_hale
}
\description{
Health-Adjusted Life Expectancy for Canada, stratified by household
income quintile. For use in Canadian DCEA analyses (CADTH workflow).
}
\keyword{datasets}
