% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_prep.R
\name{normalise_weights}
\alias{normalise_weights}
\title{Normalise population weights to sum to 1}
\usage{
normalise_weights(weights)
}
\arguments{
\item{weights}{Numeric vector of weights.}
}
\value{
Normalised numeric vector.
}
\description{
Normalise population weights to sum to 1
}
\examples{
normalise_weights(c(1, 2, 3, 4, 5))
}
