% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_dcea.R
\name{plot.aggregate_dcea}
\alias{plot.aggregate_dcea}
\title{Plot method for aggregate_dcea}
\usage{
\method{plot}{aggregate_dcea}(x, type = "impact_plane", ...)
}
\arguments{
\item{x}{An object of class \code{"aggregate_dcea"}.}

\item{type}{Plot type: \code{"impact_plane"} (default), \code{"lorenz"},
or \code{"ede_profile"}.}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Dispatches to \code{\link{plot_equity_impact_plane}} by default.
}
