% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_full_dcea.R
\name{plot.full_dcea}
\alias{plot.full_dcea}
\title{Plot method for full_dcea}
\usage{
\method{plot}{full_dcea}(x, type = "impact_plane", ...)
}
\arguments{
\item{x}{An object of class \code{"full_dcea"}.}

\item{type}{Plot type (default \code{"impact_plane"}).}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Plot method for full_dcea
}
