% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_dcea.R
\name{print.aggregate_dcea}
\alias{print.aggregate_dcea}
\title{Print method for aggregate_dcea}
\usage{
\method{print}{aggregate_dcea}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"aggregate_dcea"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for aggregate_dcea
}
