% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_full_dcea.R
\name{print.full_dcea}
\alias{print.full_dcea}
\title{Print method for full_dcea}
\usage{
\method{print}{full_dcea}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"full_dcea"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for full_dcea
}
