% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_aggregate_dcea.R
\name{summary.aggregate_dcea}
\alias{summary.aggregate_dcea}
\title{Summary method for aggregate_dcea}
\usage{
\method{summary}{aggregate_dcea}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"aggregate_dcea"}.}

\item{...}{Further arguments (ignored).}
}
\value{
A tibble of key DCEA outputs.
}
\description{
Summary method for aggregate_dcea
}
