% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_full_dcea.R
\name{summary.full_dcea}
\alias{summary.full_dcea}
\title{Summary method for full_dcea}
\usage{
\method{summary}{full_dcea}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"full_dcea"}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for full_dcea
}
